# Speed optimizations for GML and GeoJSON writing