--- a/ogr/ogr_geometry.h
+++ b/ogr/ogr_geometry.h
@@ -64,22 +64,31 @@ struct CPL_DLL OGRWktOptions
     /// Precision of output for M coordinates.  Interpretation depends on \c format.
     int mPrecision;
     /// Whether GDAL-special rounding should be applied.
-    bool round = getDefaultRound();
+    bool round;
     /// Formatting type.
     OGRWktFormat format = OGRWktFormat::Default;
 
     /// Constructor.
     OGRWktOptions()
         : xyPrecision(getDefaultPrecision()), zPrecision(xyPrecision),
-          mPrecision(zPrecision)
+          mPrecision(zPrecision), round(getDefaultRound())
+    {
+    }
+
+    /// Constructor.
+    OGRWktOptions(int xyPrecisionIn, bool roundIn)
+        : xyPrecision(xyPrecisionIn), zPrecision(xyPrecision),
+          mPrecision(zPrecision), round(roundIn)
     {
     }
 
     /// Copy constructor
     OGRWktOptions(const OGRWktOptions &) = default;
 
-  private:
+    /// Return default precision
     static int getDefaultPrecision();
+
+    /// Return default rounding mode.
     static bool getDefaultRound();
 };