Related to https://lists.osgeo.org/pipermail/gdal-dev/2024-November/059853.html

Before:
```
$ time ogr2ogr -f "mapinfo file" lakes.mif Jarvi10.shp -progress -limit 30000
real    0m7,983s

$ time ogr2ogr -f csv lakes.csv Jarvi10.shp -progress -limit 30000  -lco geometry=as_wkt
real    0m10,311s

$ time ogr2ogr -f gml lakes.gml Jarvi10.shp -progress -limit 30000
real    0m14,525s

$ time ogr2ogr -f geojson lakes.json Jarvi10.shp -progress -limit 30000
real    0m19,116s
```

After:
```
$ time ogr2ogr -f gml lakes.gml Jarvi10.shp -progress -limit 30000
real    0m12,706s

$ time ogr2ogr -f geojson lakes.json Jarvi10.shp -progress -limit 30000
real    0m15,565s
```