@rouault 2024-12-04
> are there any tests for intelliround and removeTrailingZeros ?

https://coveralls.io/builds/71175555/source?filename=ogr%2Fogrutils.cpp#L130 shows they are partially covered